requirements.txt
Django==3.1.4
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==8.3.1
PyJWT==2.1.0
translate==3.6.1
drf-yasg==1.20.0